Entire project reorganized and refactored.
