+++ src/Engine/Game.py
@@ -1,19 +1,32 @@
 import pygame as pg
 from src.Assets import settings
+import pygame
+from pygame import *
+
 
 class Game:
-
-    def __init__(self, window_width, window_height):
-        self.window = pg.display.set_mode((settings.WINDOW_WIDTH,settings.WINDOW_HEIGHT))
-        pg.display.set_caption("Gameee")
-        self.clock = pg.time.Clock()
-        self.running = True
+    def __init__(self, player, entities):
+        self.clock = pygame.time.Clock()
+        self.player = player
+        self.entities = entities
 
     def new_game(self):
-        
-    def run(self):
-        self.playing = True
-        self.clock.tick(settings.FPS)
+        pass
+
+    def run(self, window):
+        while 1:
+            for e in pygame.event.get():
+                if e.type == QUIT:
+                    return
+                if e.type == KEYDOWN and e.key == K_ESCAPE:
+                    return
+
+            self.entities.update()
+
+            window.fill((0, 0, 0))
+            self.entities.draw(window)
+            pygame.display.update()
+            self.clock.tick(settings.FPS)
